+++ src/map.rs
@@ -393,6 +393,15 @@ impl FrozenMap<Vec<u8>> {
     }
 }
 
+impl<'a, D: AsRef<[u8]>> IntoIterator for &'a FrozenMap<D> {
+    type Item = (CellIndex, u64);
+    type IntoIter = FrozenMapIterator<'a>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.iter()
+    }
+}
+
 // ------------------------------------------------------------------------------
 
 /// A builder for creating a frozen map.